Ensure that related records are updated when performing related record query
