+++ src/SIL.XForge.Scripture/ClientApp/package.json
@@ -34,10 +34,10 @@
     "@orbit/coordinator": "^0.15.23",
     "@orbit/core": "^0.15.23",
     "@orbit/data": "^0.15.23",
-    "@orbit/indexeddb": "^0.15.23",
+    "@orbit/indexeddb": "^0.15.25",
     "@orbit/indexeddb-bucket": "^0.15.23",
-    "@orbit/jsonapi": "^0.15.24",
-    "@orbit/store": "^0.15.23",
+    "@orbit/jsonapi": "^0.15.25",
+    "@orbit/store": "^0.15.25",
     "angular-oauth2-oidc": "^5.0.2",
     "angular-password-strength-meter": "0.0.4",
     "bootstrap": "^4.1.3",
